fritz/requirements.txt
fritzconnection==1.11.0
prometheus-client==0.20.0
PyYAML==6.0.2
